ui/pyproject.toml
[project]
name = "projeto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-contrib-python>=4.12.0.88",
    "opencv-python>=4.12.0.88",
    "fastapi>=0.115.2",
    "uvicorn[standard]>=0.30.6",
    "pydantic>=2.9.2",
    "pupil-apriltags>=1.0.4",
    "websockets>=15.0.1",
    "websocket-client>=1.9.0",
    "pyturbojpeg>=1.8.2",
]
